Update README.md
Tech Shop it's an open source (test scenario) software made to create a easy and simple "Shop" API, where you have two micro services, one the Products API that stores and handles everything Related to Stock and Products. And the Purchase API where you can create orders (cart's)
The purpose of this repository it's for education and test. But the code it's being coded in a proper way.